# src/shared/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: multisite-sensorpack-mvp API契約
  version: "1.0.0"
  description: |
    Web(Next.js) ⇔ Rails API ⇔ FastAPI(AIサマリー) ⇔ ESP32ファームウェア間の
    single source of truth となるAPI契約。

    本ファイルは `requirements.md` の 1.6節（関数ロジック v3・最終版）を正とし、
    その内容をOpenAPI 3.1形式のリクエスト/レスポンス形状として定義したもの。
    仕様の変更は必ず `requirements.md` を先に更新し、本ファイルへ反映すること。

    ## 対象サービスと利用者
    - `tags` に `x-service` 拡張を付与し、どのランタイムが実装するAPIかを明示する。
      - `rails-api`: Ruby on Rails（APIモード）が実装し、Next.js（Web）およびESP32が呼び出す公開API。
      - `fastapi-ai`: Python FastAPIが実装する内部AIサマリー生成API。RailsからのみServer-to-Serverで呼び出され、
        Next.js/ESP32/ブラウザから直接到達できないネットワーク配置とする（`.claude/rules/deploy.md` バックエンドドメイン隠蔽方針に準拠）。

    ## 個人情報の取り扱い（requirements.md 1.4節）
    - Googleログインの **opaqueな内部ユーザーID** のみを返却する。Google `sub` 値そのもの・メールアドレス・
      表示名・氏名・住所・電話番号・生年月日はいかなるレスポンススキーマにも含めない。
    - 拠点名（`Site.name`）は自由入力ラベルであり、住所を意味するフィールド・バリデーションは設けない。
servers:
  - url: https://sensorpack.rictaworks.jp/api/v1
    description: Rails API（本番・Next.js経由のプロキシ配下 / requirements.md 8節・deploy.mdのバックエンドドメイン隠蔽方針）
  - url: http://localhost:3000/api/v1
    description: Rails API（開発環境）
  - url: https://ai-internal.railway.internal/v1
    description: >-
      FastAPI AIサマリー内部サービス（本番）。Railwayのプライベートネットワーク内でのみ疎通し、
      公開インターネットからは到達不可能な想定のプレースホルダURL。Next.js・ESP32・ブラウザは直接呼び出さない。
  - url: http://localhost:8001/v1
    description: FastAPI AIサマリー内部サービス（開発環境）
...
security:
  - {}
paths:
...
  /alerts:
    get:
      operationId: listAlerts
      tags: [alerts]
      summary: 自分の拠点配下の全アラートを一覧取得する（UC7）
      security:
        - googleSessionCookie: []
      parameters:
        - name: status
          in: query
          description: カンマ区切りで複数指定可（例：open,acknowledged）
          schema:
            type: array
            items:
              $ref: "#/components/schemas/AlertStatus"
          style: form
          explode: false
        - name: deviceId
          in: query
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: アラート一覧
          content:
            application/json:
              schema:
                type: object
                required: [alerts]
                properties:
                  alerts:
                    type: array
                    items:
                      $ref: "#/components/schemas/Alert"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /alerts/{alertId}/ack:
    parameters:
      - name: alertId
        in: path
        required: true
        schema:
          type: integer
          format: int64
    post:
      operationId: acknowledgeAlert
      tags: [alerts]
      summary: アラートをacknowledgeする（closeは解除条件成立による自動のみ・手動close不可）
      security:
        - googleSessionCookie: []
      responses:
        "200":
          description: ack成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: 既にclosed状態のアラートはack不可
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ------------------------------------------------------------------
  # dashboard (F6)
  # ------------------------------------------------------------------
...
  /ai-summaries/today:
    get:
      operationId: getTodaySummary
      tags: [ai-summary]
      summary: 当日中に生成済みのサマリーがあれば取得する（クォータ消費なし）
      security:
        - googleSessionCookie: []
      responses:
        "200":
          description: サマリーあり
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiSummary"
        "204":
          description: 当日分は未生成
        "401":
          $ref: "#/components/responses/Unauthorized"
  /ai-summaries:
    post:
      operationId: generateDailySummary
      tags: [ai-summary]
      summary: AI日次サマリーを生成する（F7 generate_daily_summary）
      description: |
        クォータ日＝JSTの現在時刻から3時間引いた日付。同一クォータ日に既に生成済みなら429で拒否し、
        既存の保存済みサマリーを併せて返す。過去24hテレメトリ統計とアラート履歴の**統計値のみ**
        （個人情報なし）をFastAPI(internal-ai)経由でLLMに渡す。データが存在しない場合はLLMを呼ばず
        「データ不足」の定型文を返し、クォータは消費しない。
      security:
        - googleSessionCookie: []
      responses:
        "201":
          description: 新規生成成功（データ不足の定型文の場合も201・`dataSufficient=false`）
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiSummary"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          description: 当該クォータ日は既に生成済み（1日1回・JST03:00リセット相当）
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiSummaryQuotaExceeded"

  # ------------------------------------------------------------------
  # admin (F9) — BASIC認証・開発者用・日本語のみ
  # ------------------------------------------------------------------
...
components:
...
  responses:
    Unauthorized:
      description: 未認証、またはセッション/トークンが無効
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: 認証は成功しているが、他ユーザーのリソースへのアクセス（テナント分離違反）
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: リソースが存在しない、または論理削除済み
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  securitySchemes:
    googleSessionCookie:
      type: apiKey
      in: cookie
      name: session_id
      description: >-
        Googleログイン成立後にRailsが発行するhttpOnly・Secureセッションcookie。
        一般消費者が実際に使える唯一のログイン手段（開発者向け近道は本番UIに露出しない・environment.md準拠）。
    deviceBearerToken:
      type: http
      scheme: bearer
      bearerFormat: device-token
      description: >-
        ESP32デバイス登録時にclaimDeviceで発行される長寿命トークン。`Authorization: Bearer <token>`で送信する。
    internalServiceKey:
      type: apiKey
      in: header
      name: X-Internal-Api-Key
      description: Rails→FastAPIのServer-to-Server呼び出し専用の内部サービスキー。外部非公開。
    adminBasicAuth:
      type: http
      scheme: basic
      description: >-
        F9開発者用管理画面のBASIC認証。一般消費者向けGoogleログインとは別物であり混同しない
        （environment.md準拠）。
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: 機械可読なエラーコード（例：invalid_device_token, device_deleted, quota_exceeded, validation_error, forbidden, not_found）
            message:
              type: string
              description: 人間可読なエラーメッセージ（i18nキーはNext.js側で別途解決する場合はcodeを利用する）
            details:
              type: [object, "null"]
              additionalProperties: true
...
    AlertTypeCode:
      type: string
      enum: [upper_breach, lower_breach, offline]
    AlertSeverity:
      type: string
      enum: [info, warning, critical]
    AlertStatus:
      type: string
      enum: [open, acknowledged, closed]
    Alert:
      type: object
      required:
        - id
        - deviceId
        - alertType
        - severity
        - status
        - openedAt
      properties:
        id:
          type: integer
          format: int64
        deviceId:
          type: integer
          format: int64
        alertType:
          $ref: "#/components/schemas/AlertTypeCode"
        severity:
          $ref: "#/components/schemas/AlertSeverity"
        status:
          $ref: "#/components/schemas/AlertStatus"
        openedAt:
          type: string
          format: date-time
        acknowledgedAt:
          type: [string, "null"]
          format: date-time
        closedAt:
          type: [string, "null"]
          format: date-time
...
    AiSummary:
      type: object
      required: [quotaDate, summaryText, generatedAt, dataSufficient]
      properties:
        quotaDate:
          type: string
          format: date
          description: JSTの現在時刻から3時間引いた日付（JST03:00リセットと等価）
        summaryText:
          type: string
          description: 日本語の自然言語サマリー。データ不足の場合は定型文。
        generatedAt:
          type: string
          format: date-time
        dataSufficient:
          type: boolean
          description: falseの場合「データ不足」の定型文であり、LLM呼び出し・クォータ消費は発生していない
    AiSummaryQuotaExceeded:
      type: object
      required: [error, existingSummary]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum: [quota_exceeded]
            message:
              type: string
        existingSummary:
          $ref: "#/components/schemas/AiSummary"
...